requirements.txt
pygame>=2.0.0
Pillow>=9.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
requests>=2.31.0
